refactor(appkit): dedup ai-search query payload, abort bridge, and kebab-camel
Cleanups from a /simplify pass over the ai-search branch:

- _prepareQuery now returns Omit<VsQueryParams,"indexName"> (carries filters,
  renames rerankerConfig->reranker), so the /query route and programmatic
  query() collapse to `{ indexName, ...prepared }` instead of duplicating an
  8-field payload. Removes the accidental body.filters vs request.filters
  split.
- Hoist the AbortSignal->CancellationToken bridge into a shared
  connectors/context.ts (cancellationTokenFromAbortSignal +
  contextFromAbortSignal); ai-search and serving import it instead of each
  carrying a copy.
- promote.ts uses the shared kebabToCamel instead of its own
  manifestNameToBinding copy.
- _parseResponse skips the score column by index (i !== scoreIndex) rather
  than a per-cell string compare.
- Trim verbose comments introduced on the branch to the one-line bar.

Left alone (flagged, out of scope): the jobs connector's separate
_createContext variant (different listener semantics), scaffold.ts's
toCamelCase (different impl), and tools/generate-plugin-entries.ts (package
boundary).
